fix some PVS warnings and increase version to 10.1
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2469 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ flasher.cpp
@@ -439,7 +439,7 @@ void Flasher::RenderSetup(RenderDevice* pd3dDevice)
    if (dynamicVertexBuffer)
       dynamicVertexBuffer->release();
    pd3dDevice->CreateVertexBuffer(numVertices, USAGE_DYNAMIC, MY_D3DFVF_TEX, &dynamicVertexBuffer);
-   NumVideoBytes += numVertices*sizeof(Vertex3D_TexelOnly);
+   NumVideoBytes += numVertices*(int)sizeof(Vertex3D_TexelOnly);
 
    if (vertices)
       delete[] vertices;
